Added Triggers and Animation for MVP
Added in a simple box trigger, allowing the Player object to collide and teleport to a new location.
(The hit on the trigger can be changed depending on what player we have i.e one class may not be able to hit the trigger while another can)
Also added in the animation for Idle, Running, and Jumping.